优化选股分析系统：实时行情、业绩分析、信号改进
- 新增腾讯实时行情接口（GBK编码修复），补充今日K线数据
- 新增业绩分析模块（FinanceDataService），结合净利润/营收同比调整风险分
- 修复凹量洗盘bug，加入MACD金叉信号检测
- 倍量突破加阳线确认，缩量回踩加下影线承接条件
- 加入大盘联动风险（上证MA5<MA20时个股风险分+20）
- 上涨趋势+低风险无信号时给Watch而非Ignore
- 交易决策面板重构：展示趋势、操作建议、仓位、业绩数据
- 新增/name/<code>接口，支持按代码反查股票名称

+++ StockAnalysis.Core/Engines/DataImporter.cs
@@ -44,7 +44,7 @@ public class DataImporter
                 High   = csv.GetField<decimal>(2),
                 Low    = csv.GetField<decimal>(3),
                 Close  = csv.GetField<decimal>(4),
-                Volume = csv.GetField<long>(5),
+                Volume = (long)csv.GetField<double>(5),
                 Amount = csv.GetField<decimal>(6)
             });
         }
